month, day 파라메터의 타입을 u32 에서 u8 로 변경

+++ src/types/els_timetable.rs
@@ -52,7 +52,7 @@ impl ElsTimetableParams {
         self.SEM = Some(sem.to_string());
         self
     }
-    pub fn ymd(mut self, year: i32, month: u32, day: u32) -> Self {
+    pub fn ymd(mut self, year: i32, month: u8, day: u8) -> Self {
         self.ALL_TI_YMD = Some(format!("{:04}{:02}{:02}", year, month, day));
         self
     }
@@ -64,11 +64,11 @@ impl ElsTimetableParams {
         self.PERIO = Some(perio.to_string());
         self
     }
-    pub fn from_ymd(mut self, year: i32, month: u32, day: u32) -> Self {
+    pub fn from_ymd(mut self, year: i32, month: u8, day: u8) -> Self {
         self.TI_FROM_YMD = Some(format!("{:04}{:02}{:02}", year, month, day));
         self
     }
-    pub fn to_ymd(mut self, year: i32, month: u32, day: u32) -> Self {
+    pub fn to_ymd(mut self, year: i32, month: u8, day: u8) -> Self {
         self.TI_TO_YMD = Some(format!("{:04}{:02}{:02}", year, month, day));
         self
     }
